move liquid handling out of game
created a new class liquid_handler that handles liquids, added that as a member to game
