+++ src/crafting.cpp
@@ -22,6 +22,7 @@
 #include "debug.h"
 #include "game.h"
 #include "game_inventory.h"
+#include "handle_liquid.h"
 #include "inventory.h"
 #include "item.h"
 #include "item_location.h"
@@ -588,7 +589,7 @@ static item *set_item_inventory( player &p, item &newit )
 {
     item *ret_val = nullptr;
     if( newit.made_of( LIQUID ) ) {
-        g->handle_all_liquid( newit, PICKUP_RANGE );
+        g->liquid_handling.handle_all_liquid( newit, PICKUP_RANGE );
     } else {
         p.inv.assign_empty_invlet( newit, p );
         // We might not have space for the item
@@ -1068,7 +1069,7 @@ void player::complete_craft( item &craft, const tripoint &loc )
 
         finalize_crafted_item( newit );
         if( newit.made_of( LIQUID ) ) {
-            g->handle_all_liquid( newit, PICKUP_RANGE );
+            g->liquid_handling.handle_all_liquid( newit, PICKUP_RANGE );
         } else if( loc == tripoint_zero ) {
             wield_craft( *this, newit );
         } else {
@@ -1092,7 +1093,7 @@ void player::complete_craft( item &craft, const tripoint &loc )
             }
             finalize_crafted_item( bp );
             if( bp.made_of( LIQUID ) ) {
-                g->handle_all_liquid( bp, PICKUP_RANGE );
+                g->liquid_handling.handle_all_liquid( bp, PICKUP_RANGE );
             } else if( loc == tripoint_zero ) {
                 set_item_inventory( *this, bp );
             } else {
@@ -1900,7 +1901,7 @@ void player::complete_disassemble( int item_pos, const tripoint &loc,
         }
 
         if( act_item.made_of( LIQUID ) ) {
-            g->handle_all_liquid( act_item, PICKUP_RANGE );
+            g->liquid_handling.handle_all_liquid( act_item, PICKUP_RANGE );
         } else {
             drop_items.push_back( act_item );
         }
@@ -1935,7 +1936,7 @@ void remove_ammo( std::list<item> &dis_items, player &p )
 void drop_or_handle( const item &newit, player &p )
 {
     if( newit.made_of( LIQUID ) && &p == &g->u ) { // TODO: what about NPCs?
-        g->handle_all_liquid( newit, PICKUP_RANGE );
+        g->liquid_handling.handle_all_liquid( newit, PICKUP_RANGE );
     } else {
         item tmp( newit );
         p.i_add_or_drop( tmp );
