+++ src/item.cpp
@@ -33,6 +33,7 @@
 #include "game.h"
 #include "game_constants.h"
 #include "gun_mode.h"
+#include "handle_liquid.h"
 #include "iexamine.h"
 #include "item_category.h"
 #include "item_factory.h"
@@ -5150,7 +5151,7 @@ bool item::spill_contents( Character &c )
     while( !contents.empty() ) {
         on_contents_changed();
         if( contents_made_of( LIQUID ) ) {
-            if( !g->handle_liquid_from_container( *this, 1 ) ) {
+            if( !g->liquid_handling.handle_liquid_from_container( *this, 1 ) ) {
                 return false;
             }
         } else {
